+++ pegasus_sim/include/pegasus_sim/sensor_models.h
@@ -18,14 +18,14 @@ private:
   virtual void sendBarometer(const ros::TimerEvent& event);
 
   // GPS Variables
-  double nu_n_;
-  double nu_e_;
-  double nu_h_;
-  float  sigma_gps_n_;
-  float  sigma_gps_e_;
-  float  sigma_gps_h_;
-  float  sigma_gps_V_;
-  float  K_GPS_;
+  float nu_n_;
+  float nu_e_;
+  float nu_h_;
+  float sigma_gps_n_;
+  float sigma_gps_e_;
+  float sigma_gps_h_;
+  float sigma_gps_V_;
+  float K_GPS_;
 
   // Accelerometer Variables
   float mass_;
